creating search fragment

+++ app/src/main/java/com/example/weatherapp/model/geocodingmodel/Rank.kt
@@ -2,7 +2,7 @@ package com.example.weatherapp.model.geocodingmodel
 
 data class Rank(
     val confidence: Double,
-    val confidence_city_level: Int,
+    val confidence_city_level: Double,
     val importance: Double,
     val match_type: String
 )
